add: domainTag

+++ backend/src/main/java/com/a407/back/domain/Notification.java
@@ -1,22 +1,30 @@
 package com.a407.back.domain;
-import jakarta.persistence.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import java.sql.Timestamp;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.CreationTimestamp;
 
 @Table(name = "NOTIFICATION")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Entity
 public class Notification {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", updatable = false)
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "room_id")
@@ -33,7 +41,7 @@ public class Notification {
     private Timestamp createdAt;
 
     @Builder
-    public Notification(int id, String content, boolean isRead, Timestamp createdAt) {
+    public Notification(Long id, String content, boolean isRead, Timestamp createdAt) {
         this.id = id;
         this.content = content;
         this.isRead = isRead;
